requirements.txt
flask==0.12.4
Flask-Testing==0.7.1
pytest~=3.5.1
pytest-flask==0.10.0
python-dotenv~=0.8.2
flask-marshmallow==0.9.0
flask-bcrypt
marshmallow==2.15.4
marshmallow-sqlalchemy==0.14.1
flask-cors
requests
twilio~=6.15.1
rq==0.12.0
africastalking==1.1.7
nexmo==2.4.0
psycopg2-binary~=2.7.4